pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
drf-yasg = "^1.21.4"
pytest-django = "^4.5.2"
flake8 = "^5.0.4"
black = "^22.8.0"
factory-boy = "^3.2.1"
Faker = "^9.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"